Allow sampling smaller chunks
I think this is a good idea so that we don't exclude all these java files that are shorter than 512 tokens. However, we should also consider maximizing the sample size so that EOF is not favored too much. But then again, this would slightly bias towards the sequences that contain the whole short file.